bevy_image_font
Module loader

Code for parsing an ImageFont off of an on-disk representation.

Structs

ImageFontLoader

Loader for ImageFonts.

ImageFontSettings

On-disk representation of a ImageFont, optimized to make it easy for humans to write these. See the docs for ImageFontLayout’s variants for information on how to write the syntax, or [the example font’s RON asset].

Enums

ImageFontLayout

Human-readable way to specify where the characters in an image font are.

ImageFontLoadError

Errors that can show up during loading.