bevy_image_font::loader
Type Alias ImageFontSettings

pub type ImageFontSettings = ImageFontDescriptor;

👎Deprecated since 7.1.0: ImageFontSettings was renamed to ImageFontDescriptor and will be removed in version 8.0

Aliased Type

struct ImageFontSettings {
    pub image: Utf8PathBuf,
    pub layout: ImageFontLayout,
}

Fields
image: Utf8PathBuf
👎Deprecated since 7.1.0: This field will become private in the next major version. Use new to create a value of this type and the image function to read the field.

The path to the image file containing the font glyphs, relative to the RON file. This should be a valid path to a texture file that can be loaded by the asset system.

layout: ImageFontLayout
👎Deprecated since 7.1.0: This field will become private in the next major version. Use new to create a value of this type and the layout function to read the field.

The layout description of the font, specifying how characters map to regions within the image. This can use any of the variants provided by ImageFontLayout, allowing flexible configuration.