bevy_image_font::rendered
Struct ImageFontPreRenderedText

pub struct ImageFontPreRenderedText;

Available on crate feature rendered only.

A component for displaying in-world text that has been pre-rendered using an image font.

This component requires an ImageFontText component for determining its font and text. It renders its text into an image and sets it as the texture on its Sprite component.