bevy_image_font

Struct ImageFontText

#[non_exhaustive]
pub struct ImageFontText {
    pub text: String,
    pub font: Handle<ImageFont>,
    pub font_height: Option<f32>,
}

Text rendered using an ImageFont.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

text: String

The string of text to be rendered. Each character in the string is mapped to a corresponding glyph in the associated ImageFont.

font: Handle<ImageFont>

The handle to the ImageFont used to render this text. The font’s texture and atlas mapping determine how characters are displayed.

font_height: Option<f32>

If set, overrides the height the font is rendered at. This should be an integer multiple of the ‘native’ height if you want pixel accuracy, but we allow float values for things like animations.

Implementations

impl ImageFontText

pub fn text(self, value: impl Into<String>) -> Self

Sets the text field of this struct.

Examples found in repository?

examples/bmf.rs (line 87)

74fn spawn_text(mut commands: Commands, assets: Res<DemoAssets>) {
75    commands.spawn(Camera2d);
76
77    for (i, &pangram) in PANGRAMS.iter().enumerate() {
78        let font = match i % 3 {
79            0 => assets.bmf_txt.clone_weak(),
80            1 => assets.bmf_xml.clone_weak(),
81            2 => assets.bmf_bin.clone_weak(),
82            _ => unreachable!(),
83        };
84
85        commands.spawn((
86            ImageFontSpriteText::default(),
87            ImageFontText::default().text(pangram).font(font),
88            #[expect(
89                clippy::cast_precision_loss,
90                reason = "the magnitude of the numbers we're working on here are too small to lose \
91                anything"
92            )]
93            Transform::from_translation(Vec3::new(
94                0.,
95                PANGRAMS.len() as f32 * 20. + -(i as f32) * 40.,
96                0.,
97            )),
98        ));
99    }
100}

Hide additional examples

examples/rendered_ui.rs (line 75)

54fn spawn_ui(mut commands: Commands, assets: Res<DemoAssets>) {
55    commands.spawn(Camera2d);
56
57    // root node
58    commands
59        .spawn(Node {
60            width: Val::Percent(100.0),
61            height: Val::Percent(100.0),
62            align_items: AlignItems::Center,
63            justify_content: JustifyContent::Center,
64            ..default()
65        })
66        .with_children(|root| {
67            root.spawn((
68                ImageNode::default(),
69                Node {
70                    position_type: PositionType::Relative,
71                    ..default()
72                },
73                ImageFontPreRenderedUiText::default(),
74                ImageFontText::default()
75                    .text("Press SPACE to judge!")
76                    .font(assets.example.clone())
77                    .font_height(72.0),
78            ));
79        });
80
81    commands.spawn((
82        VowsNode,
83        ImageNode::default(),
84        Node {
85            position_type: PositionType::Absolute,
86            left: Val::Auto,
87            right: Val::Px(0.0),
88            ..default()
89        },
90        ImageFontText::default()
91            .font(assets.example.clone())
92            .font_height(72.0),
93    ));
94}

examples/rendered_sprite.rs (line 44)

38fn spawn_text(mut commands: Commands, assets: Res<DemoAssets>) {
39    commands.spawn(Camera2d);
40
41    commands.spawn((
42        ImageFontPreRenderedText::default(),
43        ImageFontText::default()
44            .text(TEXT)
45            .font(assets.example.clone())
46            .font_height(36.0),
47        // Our font is 45 pixels wide per character, and with an odd number of characters, the
48        // text aligns to the middle of a pixel, causing imperfect rendering. Shifting the
49        // position by 0.5 pixels ensures alignment to pixel boundaries, resulting in crisp,
50        // pixel-perfect rendering.
51        Transform::from_translation(Vec3::new(0.5, 0., 0.)),
52    ));
53    commands.spawn((
54        ImageFontPreRenderedText::default(),
55        ImageFontText::default()
56            .text(TEXT)
57            .font(assets.example.clone()),
58        // Instead of shifting the character by 0.5 pixels when the text lands in the middle of
59        // a pixel, we can anchor the sprite to an edge and move it by a whole number of pixels.
60        // To center it with the text above, we shift it left by half its width.
61        Sprite {
62            anchor: Anchor::CenterLeft,
63            ..default()
64        },
65        #[expect(
66            clippy::cast_precision_loss,
67            reason = "the magnitude of the numbers we're working on here are too small to lose \
68            anything"
69        )]
70        Transform::from_translation(Vec3::new(
71            -((TEXT.chars().count() * FONT_WIDTH / 2) as f32),
72            40.,
73            0.,
74        )),
75    ));
76}

examples/atlased_sprite.rs (line 75)

61fn setup(
62    mut commands: Commands,
63    assets: Res<DemoAssets>,
64    #[cfg(feature = "gizmos")] store: ResMut<GizmoConfigStore>,
65) {
66    #[cfg(feature = "gizmos")]
67    gizmos::configure_gizmo_defaults(store);
68
69    commands.spawn(Camera2d);
70
71    commands.spawn((
72        AnimateColor(ColorCurve::new(RAINBOW).expect("RAINBOW contains at least two colors")),
73        ImageFontSpriteText::default(),
74        ImageFontText::default()
75            .text(TEXT)
76            .font(assets.example.clone())
77            .font_height(36.0),
78        // Shift 0.5 pixels so our characters end up at integer pixel values. This is only
79        // necessary because we're using a horizontally centered anchor combined with an
80        // odd number of characters. If we used left or right alignment, we wouldn't need
81        // to do this, even with an odd number of characters.
82        Transform::from_translation(Vec3::new(0.5, 0., 0.)),
83    ));
84    commands.spawn((
85        AnimateText(TEXT, 0),
86        // This demonstrates using the `anchor` field on the `ImageFontSpriteText`.
87        // To still center it horizontally with the text above, we shift it left by half its width.
88        ImageFontSpriteText::default()
89            .color(tailwind::AMBER_500)
90            .anchor(Anchor::CenterLeft),
91        ImageFontText::default().font(assets.example.clone()),
92        #[expect(
93            clippy::cast_precision_loss,
94            reason = "the magnitude of the numbers we're working on here are too small to lose \
95            anything"
96        )]
97        Transform::from_translation(Vec3::new(
98            -((TEXT.chars().count() * FONT_WIDTH / 2) as f32),
99            40.,
100            0.,
101        )),
102    ));
103
104    commands.spawn((
105        ImageFontSpriteText::default()
106            .color(tailwind::ZINC_500)
107            .letter_spacing(LetterSpacing::Pixel(2)),
108        ImageFontText::default()
109            .text(TEXT)
110            .font(assets.example.clone())
111            .font_height(36.0),
112        Transform::from_translation(Vec3::new(0.5, -40., 0.)),
113    ));
114
115    // This will currently render without spaces; I intend to add a new feature
116    // shortly that will allow a font to have space without needing to have a
117    // space in the font image itself.
118    commands.spawn((
119        AnimateColor(ColorCurve::new(RAINBOW).expect("RAINBOW contains at least two colors")),
120        ImageFontSpriteText::default().letter_spacing(LetterSpacing::Pixel(1)),
121        ImageFontText::default()
122            .text(TEXT.to_uppercase())
123            .font(assets.variable_width.clone())
124            .font_height(32.0),
125        Transform::from_translation(Vec3::new(0., -120., 0.)),
126    ));
127}

pub fn font(self, value: impl Into<Handle<ImageFont>>) -> Self

Sets the font field of this struct.

Examples found in repository?

examples/bmf.rs (line 87)

74fn spawn_text(mut commands: Commands, assets: Res<DemoAssets>) {
75    commands.spawn(Camera2d);
76
77    for (i, &pangram) in PANGRAMS.iter().enumerate() {
78        let font = match i % 3 {
79            0 => assets.bmf_txt.clone_weak(),
80            1 => assets.bmf_xml.clone_weak(),
81            2 => assets.bmf_bin.clone_weak(),
82            _ => unreachable!(),
83        };
84
85        commands.spawn((
86            ImageFontSpriteText::default(),
87            ImageFontText::default().text(pangram).font(font),
88            #[expect(
89                clippy::cast_precision_loss,
90                reason = "the magnitude of the numbers we're working on here are too small to lose \
91                anything"
92            )]
93            Transform::from_translation(Vec3::new(
94                0.,
95                PANGRAMS.len() as f32 * 20. + -(i as f32) * 40.,
96                0.,
97            )),
98        ));
99    }
100}

Hide additional examples

examples/rendered_ui.rs (line 76)

54fn spawn_ui(mut commands: Commands, assets: Res<DemoAssets>) {
55    commands.spawn(Camera2d);
56
57    // root node
58    commands
59        .spawn(Node {
60            width: Val::Percent(100.0),
61            height: Val::Percent(100.0),
62            align_items: AlignItems::Center,
63            justify_content: JustifyContent::Center,
64            ..default()
65        })
66        .with_children(|root| {
67            root.spawn((
68                ImageNode::default(),
69                Node {
70                    position_type: PositionType::Relative,
71                    ..default()
72                },
73                ImageFontPreRenderedUiText::default(),
74                ImageFontText::default()
75                    .text("Press SPACE to judge!")
76                    .font(assets.example.clone())
77                    .font_height(72.0),
78            ));
79        });
80
81    commands.spawn((
82        VowsNode,
83        ImageNode::default(),
84        Node {
85            position_type: PositionType::Absolute,
86            left: Val::Auto,
87            right: Val::Px(0.0),
88            ..default()
89        },
90        ImageFontText::default()
91            .font(assets.example.clone())
92            .font_height(72.0),
93    ));
94}

examples/rendered_sprite.rs (line 45)

38fn spawn_text(mut commands: Commands, assets: Res<DemoAssets>) {
39    commands.spawn(Camera2d);
40
41    commands.spawn((
42        ImageFontPreRenderedText::default(),
43        ImageFontText::default()
44            .text(TEXT)
45            .font(assets.example.clone())
46            .font_height(36.0),
47        // Our font is 45 pixels wide per character, and with an odd number of characters, the
48        // text aligns to the middle of a pixel, causing imperfect rendering. Shifting the
49        // position by 0.5 pixels ensures alignment to pixel boundaries, resulting in crisp,
50        // pixel-perfect rendering.
51        Transform::from_translation(Vec3::new(0.5, 0., 0.)),
52    ));
53    commands.spawn((
54        ImageFontPreRenderedText::default(),
55        ImageFontText::default()
56            .text(TEXT)
57            .font(assets.example.clone()),
58        // Instead of shifting the character by 0.5 pixels when the text lands in the middle of
59        // a pixel, we can anchor the sprite to an edge and move it by a whole number of pixels.
60        // To center it with the text above, we shift it left by half its width.
61        Sprite {
62            anchor: Anchor::CenterLeft,
63            ..default()
64        },
65        #[expect(
66            clippy::cast_precision_loss,
67            reason = "the magnitude of the numbers we're working on here are too small to lose \
68            anything"
69        )]
70        Transform::from_translation(Vec3::new(
71            -((TEXT.chars().count() * FONT_WIDTH / 2) as f32),
72            40.,
73            0.,
74        )),
75    ));
76}

examples/atlased_sprite.rs (line 76)

61fn setup(
62    mut commands: Commands,
63    assets: Res<DemoAssets>,
64    #[cfg(feature = "gizmos")] store: ResMut<GizmoConfigStore>,
65) {
66    #[cfg(feature = "gizmos")]
67    gizmos::configure_gizmo_defaults(store);
68
69    commands.spawn(Camera2d);
70
71    commands.spawn((
72        AnimateColor(ColorCurve::new(RAINBOW).expect("RAINBOW contains at least two colors")),
73        ImageFontSpriteText::default(),
74        ImageFontText::default()
75            .text(TEXT)
76            .font(assets.example.clone())
77            .font_height(36.0),
78        // Shift 0.5 pixels so our characters end up at integer pixel values. This is only
79        // necessary because we're using a horizontally centered anchor combined with an
80        // odd number of characters. If we used left or right alignment, we wouldn't need
81        // to do this, even with an odd number of characters.
82        Transform::from_translation(Vec3::new(0.5, 0., 0.)),
83    ));
84    commands.spawn((
85        AnimateText(TEXT, 0),
86        // This demonstrates using the `anchor` field on the `ImageFontSpriteText`.
87        // To still center it horizontally with the text above, we shift it left by half its width.
88        ImageFontSpriteText::default()
89            .color(tailwind::AMBER_500)
90            .anchor(Anchor::CenterLeft),
91        ImageFontText::default().font(assets.example.clone()),
92        #[expect(
93            clippy::cast_precision_loss,
94            reason = "the magnitude of the numbers we're working on here are too small to lose \
95            anything"
96        )]
97        Transform::from_translation(Vec3::new(
98            -((TEXT.chars().count() * FONT_WIDTH / 2) as f32),
99            40.,
100            0.,
101        )),
102    ));
103
104    commands.spawn((
105        ImageFontSpriteText::default()
106            .color(tailwind::ZINC_500)
107            .letter_spacing(LetterSpacing::Pixel(2)),
108        ImageFontText::default()
109            .text(TEXT)
110            .font(assets.example.clone())
111            .font_height(36.0),
112        Transform::from_translation(Vec3::new(0.5, -40., 0.)),
113    ));
114
115    // This will currently render without spaces; I intend to add a new feature
116    // shortly that will allow a font to have space without needing to have a
117    // space in the font image itself.
118    commands.spawn((
119        AnimateColor(ColorCurve::new(RAINBOW).expect("RAINBOW contains at least two colors")),
120        ImageFontSpriteText::default().letter_spacing(LetterSpacing::Pixel(1)),
121        ImageFontText::default()
122            .text(TEXT.to_uppercase())
123            .font(assets.variable_width.clone())
124            .font_height(32.0),
125        Transform::from_translation(Vec3::new(0., -120., 0.)),
126    ));
127}

pub fn font_height(self, value: impl Into<Option<f32>>) -> Self

Sets the font_height field of this struct.

Examples found in repository?

examples/rendered_ui.rs (line 77)

54fn spawn_ui(mut commands: Commands, assets: Res<DemoAssets>) {
55    commands.spawn(Camera2d);
56
57    // root node
58    commands
59        .spawn(Node {
60            width: Val::Percent(100.0),
61            height: Val::Percent(100.0),
62            align_items: AlignItems::Center,
63            justify_content: JustifyContent::Center,
64            ..default()
65        })
66        .with_children(|root| {
67            root.spawn((
68                ImageNode::default(),
69                Node {
70                    position_type: PositionType::Relative,
71                    ..default()
72                },
73                ImageFontPreRenderedUiText::default(),
74                ImageFontText::default()
75                    .text("Press SPACE to judge!")
76                    .font(assets.example.clone())
77                    .font_height(72.0),
78            ));
79        });
80
81    commands.spawn((
82        VowsNode,
83        ImageNode::default(),
84        Node {
85            position_type: PositionType::Absolute,
86            left: Val::Auto,
87            right: Val::Px(0.0),
88            ..default()
89        },
90        ImageFontText::default()
91            .font(assets.example.clone())
92            .font_height(72.0),
93    ));
94}

Hide additional examples

examples/rendered_sprite.rs (line 46)

38fn spawn_text(mut commands: Commands, assets: Res<DemoAssets>) {
39    commands.spawn(Camera2d);
40
41    commands.spawn((
42        ImageFontPreRenderedText::default(),
43        ImageFontText::default()
44            .text(TEXT)
45            .font(assets.example.clone())
46            .font_height(36.0),
47        // Our font is 45 pixels wide per character, and with an odd number of characters, the
48        // text aligns to the middle of a pixel, causing imperfect rendering. Shifting the
49        // position by 0.5 pixels ensures alignment to pixel boundaries, resulting in crisp,
50        // pixel-perfect rendering.
51        Transform::from_translation(Vec3::new(0.5, 0., 0.)),
52    ));
53    commands.spawn((
54        ImageFontPreRenderedText::default(),
55        ImageFontText::default()
56            .text(TEXT)
57            .font(assets.example.clone()),
58        // Instead of shifting the character by 0.5 pixels when the text lands in the middle of
59        // a pixel, we can anchor the sprite to an edge and move it by a whole number of pixels.
60        // To center it with the text above, we shift it left by half its width.
61        Sprite {
62            anchor: Anchor::CenterLeft,
63            ..default()
64        },
65        #[expect(
66            clippy::cast_precision_loss,
67            reason = "the magnitude of the numbers we're working on here are too small to lose \
68            anything"
69        )]
70        Transform::from_translation(Vec3::new(
71            -((TEXT.chars().count() * FONT_WIDTH / 2) as f32),
72            40.,
73            0.,
74        )),
75    ));
76}

examples/atlased_sprite.rs (line 77)

61fn setup(
62    mut commands: Commands,
63    assets: Res<DemoAssets>,
64    #[cfg(feature = "gizmos")] store: ResMut<GizmoConfigStore>,
65) {
66    #[cfg(feature = "gizmos")]
67    gizmos::configure_gizmo_defaults(store);
68
69    commands.spawn(Camera2d);
70
71    commands.spawn((
72        AnimateColor(ColorCurve::new(RAINBOW).expect("RAINBOW contains at least two colors")),
73        ImageFontSpriteText::default(),
74        ImageFontText::default()
75            .text(TEXT)
76            .font(assets.example.clone())
77            .font_height(36.0),
78        // Shift 0.5 pixels so our characters end up at integer pixel values. This is only
79        // necessary because we're using a horizontally centered anchor combined with an
80        // odd number of characters. If we used left or right alignment, we wouldn't need
81        // to do this, even with an odd number of characters.
82        Transform::from_translation(Vec3::new(0.5, 0., 0.)),
83    ));
84    commands.spawn((
85        AnimateText(TEXT, 0),
86        // This demonstrates using the `anchor` field on the `ImageFontSpriteText`.
87        // To still center it horizontally with the text above, we shift it left by half its width.
88        ImageFontSpriteText::default()
89            .color(tailwind::AMBER_500)
90            .anchor(Anchor::CenterLeft),
91        ImageFontText::default().font(assets.example.clone()),
92        #[expect(
93            clippy::cast_precision_loss,
94            reason = "the magnitude of the numbers we're working on here are too small to lose \
95            anything"
96        )]
97        Transform::from_translation(Vec3::new(
98            -((TEXT.chars().count() * FONT_WIDTH / 2) as f32),
99            40.,
100            0.,
101        )),
102    ));
103
104    commands.spawn((
105        ImageFontSpriteText::default()
106            .color(tailwind::ZINC_500)
107            .letter_spacing(LetterSpacing::Pixel(2)),
108        ImageFontText::default()
109            .text(TEXT)
110            .font(assets.example.clone())
111            .font_height(36.0),
112        Transform::from_translation(Vec3::new(0.5, -40., 0.)),
113    ));
114
115    // This will currently render without spaces; I intend to add a new feature
116    // shortly that will allow a font to have space without needing to have a
117    // space in the font image itself.
118    commands.spawn((
119        AnimateColor(ColorCurve::new(RAINBOW).expect("RAINBOW contains at least two colors")),
120        ImageFontSpriteText::default().letter_spacing(LetterSpacing::Pixel(1)),
121        ImageFontText::default()
122            .text(TEXT.to_uppercase())
123            .font(assets.variable_width.clone())
124            .font_height(32.0),
125        Transform::from_translation(Vec3::new(0., -120., 0.)),
126    ));
127}