bevy_image_font::atlas_sprites
Function set_up_sprites

pub fn set_up_sprites(
    commands: Commands<'_, '_>,
    query: Query<'_, '_, (Entity, &ImageFontText, &ImageFontSpriteText, Option<&mut ImageFontTextData>), Or<(Changed<ImageFontText>, Changed<ImageFontSpriteText>)>>,
    child_query: Query<'_, '_, (&mut Sprite, &mut Transform)>,
    image_fonts: Res<'_, Assets<ImageFont>>,
    texture_atlas_layouts: Res<'_, Assets<TextureAtlasLayout>>,
)

Available on crate feature atlas_sprites only.

System that renders each ImageFontText as child [Sprite] entities where each sprite represents a character in the text. That is to say, each sprite gets positioned accordingly to its position in the text. This system only runs when the ImageFontText or ImageFontSpriteText changes.