bevy_image_font

Module loader

Code for parsing an ImageFont off of an on-disk representation.

Structs

BmFontLoaderbmf

Loader for ImageFonts.

ImageFontDescriptor

On-disk representation of an ImageFont, optimized to make it easy for humans to write these. See the docs for ImageFontLayout’s variants for information on how to write the syntax, or [the example font’s RON asset].

ImageFontLoader

Loader for ImageFonts.

ImageFontLoaderSettings

Configuration settings for the ImageFontLoader.

Enums

BmFontLoadErrorbmf

An error type representing issues that may arise during the loading of BMF fonts.

ImageFontDescriptorValidationError

Errors that can show up during validation.

ImageFontLayout

Human-readable way to specify where the characters in an image font are.

ImageFontLayoutValidationError

Errors that can show up during validation.

ImageFontLoadError

Errors that can show up during font loading.